feat: Add name and schoolName fields to CreateUploadDto and Upload entity
This commit adds the `name` and `schoolName` fields to the `CreateUploadDto` class in the `upload/dto/create-upload.dto.ts` file. It also adds the corresponding fields to the `Upload` entity in the `upload/entities/upload.entity.ts` file.

The `UploadService` in the `upload/upload.service.ts` file has been updated to include the `name` and `schoolName` fields when retrieving uploads.

In the `frontend/src/components/tracking/wish.component.jsx` file, the table headers for the wish component have been modified to include the "Họ tên" (name) and "Trường" (school) columns. The table rows have also been updated to display the `name` and `schoolName` values for each wish.

Lastly, in the `frontend/src/main.jsx` file, the `ProtectedRoute` element has been updated to use the `WishPage` component instead of the `TrackingPage` component.